<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

<!-- Describe the purpose and goals of this pull request. -->
update deepep version to support `Qwen/Qwen3-235B-A22B-FP8` , refer https://github.com/deepseek-ai/DeepEP/pull/329
## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->
```
python3 -m sglang.launch_server --model-path Qwen/Qwen3-Coder-480B-A35B-Instruct-FP8 --host 0.0.0.0 --port 40000 --trust-remote-code  --tp-size 8 --moe-a2a-backend deepep --mem-fraction-static 0.8
python3 benchmark/gsm8k/bench_sglang.py --port 40000 
Accuracy: 0.955
Invalid: 0.000
Latency: 37.781 s
Output throughput: 709.087 token/s
```
## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->
```
python3 -m sglang.launch_server --model-path Qwen/Qwen3-235B-A22B-FP8  --host 0.0.0.0 --port 40000 --trust-remote-code  --tp-size 8 --dp 8 --moe-dense-tp-size 1 --enable-dp-lm-head --moe-a2a-backend deepep --mem-fraction-static 0.8 --enable-dp-attention --max-running-requests 1024 --deepep-mode auto --cuda-graph-max-bs 128

# before deepep version

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |             63.256 |              63.256 |        256.467 |          254.349 |       263.485 |         15.563 |           15.576 |        15.589 |                63.256 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            218.609 |             218.609 |        573.219 |          573.862 |       715.690 |         17.733 |           17.668 |        18.273 |                54.652 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |            723.531 |             723.531 |        708.242 |          707.572 |       917.711 |         21.418 |           21.516 |        22.234 |                45.221 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           1322.587 |            1322.587 |        942.452 |          925.061 |      1381.734 |         23.267 |           23.279 |        24.062 |                41.331 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
# after deepep version

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |             63.341 |              63.341 |        240.267 |          235.994 |       251.449 |         15.559 |           15.557 |        15.571 |                63.341 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            218.415 |             218.415 |        576.771 |          571.467 |       791.002 |         17.746 |           17.623 |        18.327 |                54.604 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |            724.131 |             724.131 |        693.155 |          710.137 |       945.936 |         21.416 |           21.460 |        22.266 |                45.258 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           1321.885 |            1321.885 |        895.186 |          896.036 |      1343.294 |         23.328 |           23.364 |        24.231 |                41.309 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```
## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).