@yizhang2077 2025-08-11
> Could you update the CI based on gemini's suggestion?

done